+++ src/Http/AdminApi/ErrorHandling/Exception/AbstractRequestException.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Package\Shopware6\Http\AdminApi\ErrorHandling\Exception;
+
+use Psr\Http\Client\RequestExceptionInterface;
+use Psr\Http\Message\RequestInterface;
+
+abstract class AbstractRequestException extends \RuntimeException implements RequestExceptionInterface
+{
+    protected RequestInterface $request;
+
+    public function __construct(RequestInterface $request, ?string $message = '', int $code = 0, ?\Throwable $previous = null)
+    {
+        parent::__construct($message, $code, $previous);
+        $this->request = $request;
+    }
+
+    public function getRequest(): RequestInterface
+    {
+        return $this->request;
+    }
+}